삭제 바로 적용 안됨/ 수정기능 자체는 가능하나 다른 홈페이지로 이동하는 기능이 안됨.

+++ src/store_post.ts
@@ -76,17 +76,17 @@ async function setPostData() {
         })
     }
     const removePost = async (removepost: Post) => {
-        const deleteResult = await axios.delete(
-            `https://example-crud-api-using-next-jihoon416.vercel.app/api/${member}/delete/${removepost.id}`,
-        )
-        console.log('안녕')
-        update((datas) => {
-            const setData = postLists.filter(
-                (post) => post.id !== deleteResult.data.post.id,
+        if (window.confirm('정말로 삭제하시겠습니까?')) {
+            const deleteResult = await axios.delete(
+                `https://example-crud-api-using-next-jihoon416.vercel.app/api/${member}/delete/${removepost.id}`,
             )
-            datas = setData
-            return datas
-        })
+            update((datas) => {
+                const setData = postLists.filter(
+                    (post) => post !== deleteResult.data.post,
+                )
+                return setData
+            })
+        }
     }
 
     return {
